Update of files

+++ docs/release/src/emu/digfx.h
@@ -23,10 +23,10 @@
 //  CONSTANTS
 //**************************************************************************
 
-const int MAX_GFX_ELEMENTS = 32;
-const int MAX_GFX_PLANES = 8;
+const UINT8 MAX_GFX_ELEMENTS = 32;
+const UINT16 MAX_GFX_PLANES = 8;
 // HBMAME - needed by monaco
-const int MAX_GFX_SIZE = 64;
+const UINT16 MAX_GFX_SIZE = 64;
 
 
 
@@ -202,13 +202,13 @@ public:
 
 	// getters
 	palette_device &palette() const { assert(m_palette != nullptr); return *m_palette; }
-	gfx_element *gfx(int index) const { assert(index < MAX_GFX_ELEMENTS); return m_gfx[index].get(); }
+	gfx_element *gfx(UINT8 index) const { assert(index < MAX_GFX_ELEMENTS); return m_gfx[index].get(); }
 
 	// decoding
 	void decode_gfx(const gfx_decode_entry *gfxdecodeinfo);
 	void decode_gfx() { decode_gfx(m_gfxdecodeinfo); }
 
-	void set_gfx(int index, std::unique_ptr<gfx_element> &&element) { assert(index < MAX_GFX_ELEMENTS); m_gfx[index] = std::move(element); }
+	void set_gfx(UINT8 index, std::unique_ptr<gfx_element> &&element) { assert(index < MAX_GFX_ELEMENTS); m_gfx[index] = std::move(element); }
 
 protected:
 	// interface-level overrides
